pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["faroukibrahim-fii <farouk94i@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
djangorestframework = "^3.13.0"
pytz = "^2021.3"
psycopg2-binary = "^2.9.2"


[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
